refactor(web): split API into multiple files
