+++ web/src/api/index.ts
@@ -1,0 +1,5 @@
+export * from "./auth"
+export * from "./execution"
+export * from "./player"
+export * from "./templates"
+export * from "./utils"
