+++ web/src/routes/login.tsx
@@ -6,12 +6,12 @@ import {
 } from "react-router-dom"
 import { Button, Collapse, Form as FormBS } from "react-bootstrap"
 import { tryFetchApi } from "../api"
-import { isLoggedIn } from "../utils"
-import { setStorageItem } from "../storage"
+import { setStorageItem } from "../services/storage"
 import { ReactElement } from "react"
 import { CsrfForm } from "../components/CsrfForm"
 
 import "./login.css"
+import { isLoggedIn } from "../services/auth"
 
 export function LoginRoute(): ReactElement {
   const error = useActionData() as string
