--- web/src/utils.ts
@@ -1,3 +1,0 @@
-export function isLoggedIn(): boolean {
-  return localStorage.getItem("token") !== null
-}
